fix: omit task repositories on lookup failure
